refactor: make the stepper maintenable
