+++ web/beacon-app/src/components/layout/Sidebar/index.ts
@@ -1,2 +1,2 @@
-export { default as OnBoardingSidebar } from './onboard-sidebar';
+export { default as OnBoardingSidebar } from './OnboardingSideBar';
 export { default as Sidebar } from './Sidebar';
