--- web/beacon-app/src/features/onboarding/components/Stepper.tsx
@@ -1,9 +1,0 @@
-const Stepper = () => {
-  return (
-    <div>
-      <h1>Stepper</h1>
-    </div>
-  );
-};
-
-export default Stepper;
