+++ web/beacon-app/src/features/onboarding/components/stepper/Indicator.tsx
@@ -1,0 +1,9 @@
+// isolate the indicator
+// TODO: add active state to the indicator
+const Indicator = () => {
+  return (
+    <span className="absolute -left-[4px] mt-1 flex h-2 w-2 items-center justify-center rounded-full bg-gray-100 ring-4 ring-white"></span>
+  );
+};
+
+export default Indicator;
