+++ web/beacon-app/src/features/onboarding/components/stepper/index.tsx
@@ -1,0 +1,1 @@
+export { default as Stepper } from './Stepper';
