+++ web/beacon-app/src/features/onboarding/routes/OnboardingPage.tsx
@@ -1,12 +1,10 @@
 import AppLayout from '@/components/layout/AppLayout';
 
 import Step from '../components/Step';
-import Stepper from '../components/Stepper';
 
 const OnboardingPage = () => {
   return (
     <AppLayout>
-      <Stepper />
       <Step />
     </AppLayout>
   );
